Improving type hinting

+++ src/v3xctrl_telemetry/apps/read_battery.py
@@ -1,9 +1,9 @@
 from v3xctrl_telemetry import Battery
 import time
 
-min_voltage = 3.5 * 1000
-max_voltage = 4.2 * 1000
-warn_voltage = 3.7 * 1000
+min_voltage = int(3.5 * 1000)
+max_voltage = int(4.2 * 1000)
+warn_voltage = int(3.7 * 1000)
 
 battery = Battery(min_voltage, max_voltage, warn_voltage)
